requirements.txt
streamlit
geocoder
textblob
groq
python-dotenv
nltk